requirements.txt
PyQt6>=6.4.0
openai-whisper>=20240930
torch>=2.0.0
numpy>=1.24.0
psutil>=5.9.0
